requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
# Raspberry Pi dependencies
opencv-python>=4.5.0
pyserial>=3.5
requests>=2.25.0
numpy>=1.19.0

